Avoid vector default-init performance bottleneck
The lookup map in assemble_pre and assemble_post does not need to be
zero'd, as all used values are set upon building it. As it needs to be
of size n to be able to map any row index, zeroing it can actually be
quite costly, especially since it is repeated for each leaf. Very
noticeable for example on DIMACS10/tx2010 from SuiteSparse matrix
collection.

This replaces the std::vector (which always default initializes to 0)
with std::unique_ptr to avoid the repeated zeroing.