Allow small leaf subtree with only one node
"No point for a single node" means that for a single node, exactly one
task would be run anyways. So there is nothing to amalgamated into a
single task, as seems to be the idea of the small leaf subtrees. But
perhaps there could be other optimizations enabled for small leafs, so
let's try without the restriction.